Functions

Functions

Main Functions

SyntheticGrids.place_loads_from_zips! — Function.

place_loads_from_zips!(
    grid::Grid;
    latlim = (-Inf,Inf),
    longlim = (-Inf,Inf),
    datapath = CENSUSPATH,
    loadfunc = linearload,
    voltfunc = randvolt
)

Create load buses based on zipcode information.

Arguments

grid: The synthetic grid instance that will be altered.

latlim: Tuple of the form (minimum latitude, maximum latitude).

longlim: Tuple of the form (minimum longitude, maximum longitude).

datapath: Path to the text file containing census information.

loadfunc: Function for computing bus load based on some rule.

voltfunc: Function for computing bus voltage based on some rule.

REFERENCE: https://www.census.gov/geo/maps-data/data/gazetteer2010.html

source

place_loads_from_zips!(
    grid::Grid,
    geolim::Function;
    datapath = CENSUSPATH,
    loadfunc = linearload,
    voltfunc = randvolt
)

Create load buses based on zipcode information.

Arguments

grid: The synthetic grid instance that will be altered.

geolim: Function that receives a latitude-longitude pair and returns true or false.

Specifies the region of interest.

datapath: Path to the text file containing census information.

loadfunc: Function for computing bus load based on some rule.

voltfunc: Function for computing bus voltage based on some rule.

REFERENCE: https://www.census.gov/geo/maps-data/data/gazetteer2010.html

source

SyntheticGrids.place_gens_from_data! — Function.

place_gens_from_data!(
    grid::Grid,
    plants::Array,
    pfactorfunc::Function = pfacwavg
)

Create generation buses based on power plant data.

Arguments:

grid: Grid instance which will be populated.

plants: Power plant information already parsed.

pfactorfunc: Function for computing bus power factor based on some rule.

source

place_gens_from_data!(
    grid::Grid;
    latlim::Tuple = (-Inf, Inf),
    longlim::Tuple = (-Inf, Inf),
    jsonpath = GENJSONPATH,
    pfactorfunc::Function = pfacwavg,
)

Create generation buses based on power plant data.

Arguments:

grid: Grid instance which will be populated.

latlim: Tuple of the form (minimum latitude, maximum latitude).

longlim: Tuple of the form (minimum longitude, maximum longitude).

jsonpath: Path to the JSON file with the generator data.

pfactorfunc: Function for computing bus power factor based on some rule.

REFERENCE: https://www.eia.gov/electricity/data/eia860/index.html

source

place_gens_from_data!(
    grid::Grid,
    geolim::Function;
    jsonpath = GENJSONPATH,
    pfactorfunc::Function = pfacwavg,
)

Create generation buses based on power plant data.

Arguments:

grid: Grid instance which will be populated.

geolim: Function that receives a latitude-longitude pair and returns true or false.

Specifies the region of interest.

jsonpath: Path to the JSON file with the generator data.

pfactorfunc: Function for computing bus power factor based on some rule.

REFERENCE: https://www.eia.gov/electricity/data/eia860/index.html

source

SyntheticGrids.connect! — Function.

connect!(grid::Grid; k=2.5, m=-1, a=1, b=3.2, g=2.5, t=2, N=10)

Connect buses in an electric grid.

Arguments:

k: weight of the spanning tree (see REFERENCE)

m: total number of connections (computed from the # of nodes if default)

a, b, g, t: paramaters of the model (see REFERENCE)

N: number of nearest neighbors in the average distance computation

REFERENCE: Soltan, Saleh, and Gil Zussman. "Generation of synthetic spatially embedded power grid networks." Power and Energy Society General Meeting (PESGM), 2016. IEEE, 2016.

source

SyntheticGrids.cluster! — Function.

cluster!(grid::Grid, nloads, nboth, ngens)

Cluster all nodes of a grid into 'nloads' load substations, 'ngens' generation substations and 'nboth' substations with both load and generation.

REFERENCE: Birchfield, Adam B., et al. "Grid Structural Characteristics as Validation Criteria for Synthetic Networks." IEEE Transactions on Power Systems (2016).

source

SyntheticGrids.create_lines! — Function.

create_lines!(grid::Grid; impedfunc = linear_imped, capfunc = volt_cap)

Create transmission lines for a synthetic grid. Line impedancies are determined by impedfunc and line capacities are calculated via capfunc. This function will only work after the grid has been connected through the connect!() function.

source

SyntheticGrids.add_load! — Function.

add_load!(grid::Grid, args...; reconnect = false)

Add LoadBus to grid by calling the LoadBus(coords, load, volt, pop, connected_to = Set(), connections = Set()) method with args... as arguments. If reconnect = true, all connections will be remade.

source

SyntheticGrids.add_gen! — Function.

add_gen!(grid::Grid, args...; reconnect = false)

Add Genbus to grid by calling the GenBus(coords, gen, volt, tech, connected_to = Set(), connections = Set(), pfactor = -1, summgen = -1, wintgen = -1, gens = []) method with args... as arguments. If reconnect = true, all connections will be remade.

source

SyntheticGrids.add_substation! — Function.

add_substation!(grid::Grid, args...; reconnect = false)

Add Substation to grid by calling the Substation(coords, volts, load, gen, pop, con = Set(), group = []) method with args... as arguments.

source

Export/Import Functions

to_pandapower

save_pp_grid

load_pp_grid

Topology Building Functions

TWST!

reinforce!

connect_buses!

connect_subs!

cluster_loads!

cluster_load_gen!

cluster_gens!

Checking Functions

total_links

mean_node_deg

cluster_coeff

cluster_coeff_degw

test_connectivity

mean_short_path_hop

mean_short_path_real_bus

mean_short_path_real_sub

robustness_line

robustness_node